Use exists() and has() correctly
exists() and has() are two different functions that test different
things.

exists() is used for options. see `:help exists`, but use
`exists('+option')` to tell if an option exists vs. `exists('option')`
to tell if the option is set.

has() is used to tell if certain features are compiled into the current
instance.