zsh: append_strings() was messing up zle
The append_strings() function in .zprofile had a trap that was
preventing a login zsh from accepting CTRL-C correctly, most notably by
not being able to cancel a current command edit.

The new function sucks and doesn't actually format correctly. Seems to
be something weird with the zsh printf space padding functionality. It
works in a local script form :-/